+++ tech1-framework-domain/src/test/java/io/tech1/framework/domain/properties/utilties/PropertiesAsserterAndPrinterTest.java
@@ -97,7 +97,7 @@ public class PropertiesAsserterAndPrinterTest {
         var emailConfigs = EmailConfigs.of(
                 false,
                 null,
-                null,
+                0,
                 null,
                 null,
                 null,