+++ tech1-framework-emails/src/main/java/io/tech1/framework/emails/services/impl/EmailServiceImpl.java
@@ -1,22 +1,29 @@
 package io.tech1.framework.emails.services.impl;
 
+import io.tech1.framework.emails.domain.EmailHTML;
 import io.tech1.framework.emails.services.EmailService;
+import io.tech1.framework.emails.utilities.EmailUtility;
 import io.tech1.framework.properties.ApplicationFrameworkProperties;
 import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.mail.SimpleMailMessage;
 import org.springframework.mail.javamail.JavaMailSender;
+import org.thymeleaf.context.Context;
+import org.thymeleaf.spring5.SpringTemplateEngine;
 
+import javax.mail.MessagingException;
 import java.util.List;
 import java.util.Set;
 
-@Slf4j
 @RequiredArgsConstructor(onConstructor = @__(@Autowired))
 public class EmailServiceImpl implements EmailService {
 
     // Services
     private final JavaMailSender javaMailSender;
+    // HTML Engine
+    private final SpringTemplateEngine springTemplateEngine;
+    // Utilities
+    private final EmailUtility emailUtility;
     // Properties
     private final ApplicationFrameworkProperties applicationFrameworkProperties;
 
@@ -30,9 +37,8 @@ public class EmailServiceImpl implements EmailService {
     public void sendPlain(String[] to, String subject, String message) {
         var emailConfigs = this.applicationFrameworkProperties.getEmailConfigs();
         if (emailConfigs.isEnabled()) {
-            var from = emailConfigs.getFrom();
             var mailMessage = new SimpleMailMessage();
-            mailMessage.setFrom(from);
+            mailMessage.setFrom(emailConfigs.getFrom());
             mailMessage.setTo(to);
             mailMessage.setSubject(subject);
             mailMessage.setText(message);
@@ -48,5 +54,27 @@ public class EmailServiceImpl implements EmailService {
     @Override
     public void sendPlain(Set<String> to, String subject, String message) {
         this.sendPlain(to.toArray(new String[0]), subject, message);
+    }
+
+    @Override
+    public void sendHTML(EmailHTML emailHTML) {
+        var emailConfigs = this.applicationFrameworkProperties.getEmailConfigs();
+        if (emailConfigs.isEnabled()) {
+            try {
+                var tuple2 = this.emailUtility.getMimeMessageTuple2();
+                var message = tuple2.getA();
+                var mmHelper = tuple2.getB();
+                mmHelper.setFrom(emailConfigs.getFrom());
+                mmHelper.setTo(emailHTML.getTo().toArray(new String[0]));
+                mmHelper.setSubject(emailHTML.getSubject());
+                var context = new Context();
+                context.setVariables(emailHTML.getTemplateVariables());
+                var processedHTML = this.springTemplateEngine.process(emailHTML.getTemplateName(), context);
+                mmHelper.setText(processedHTML, true);
+                this.javaMailSender.send(message);
+            } catch (MessagingException ex) {
+                // ignored
+            }
+        }
     }
 }