+++ tech1-framework-emails/src/test/java/io/tech1/framework/emails/services/impl/EmailServiceImplTest.java
@@ -1,7 +1,10 @@
 package io.tech1.framework.emails.services.impl;
 
 import io.tech1.framework.domain.properties.configs.EmailConfigs;
+import io.tech1.framework.domain.tuples.Tuple2;
+import io.tech1.framework.emails.domain.EmailHTML;
 import io.tech1.framework.emails.services.EmailService;
+import io.tech1.framework.emails.utilities.EmailUtility;
 import io.tech1.framework.properties.ApplicationFrameworkProperties;
 import lombok.RequiredArgsConstructor;
 import org.junit.jupiter.api.AfterEach;
@@ -14,13 +17,22 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.mail.SimpleMailMessage;
 import org.springframework.mail.javamail.JavaMailSender;
+import org.springframework.mail.javamail.MimeMessageHelper;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
 import org.springframework.test.context.support.AnnotationConfigContextLoader;
+import org.thymeleaf.spring5.SpringTemplateEngine;
+import org.thymeleaf.spring5.templateresolver.SpringResourceTemplateResolver;
+import org.thymeleaf.templatemode.TemplateMode;
 
+import javax.mail.MessagingException;
+import javax.mail.internet.MimeMessage;
+import java.nio.charset.StandardCharsets;
 import java.util.List;
+import java.util.Map;
 import java.util.Set;
 
+import static io.tech1.framework.domain.utilities.random.EntityUtility.entity;
 import static io.tech1.framework.domain.utilities.random.RandomUtility.randomEmail;
 import static io.tech1.framework.domain.utilities.random.RandomUtility.randomString;
 import static org.assertj.core.api.Assertions.assertThat;
@@ -44,9 +56,33 @@ public class EmailServiceImplTest {
         }
 
         @Bean
+        SpringTemplateEngine springTemplateEngine() {
+            var templateEngine = new SpringTemplateEngine();
+            templateEngine.addTemplateResolver(htmlTemplateResolver());
+            return templateEngine;
+        }
+
+        @Bean
+        SpringResourceTemplateResolver htmlTemplateResolver() {
+            var emailTemplateResolver = new SpringResourceTemplateResolver();
+            emailTemplateResolver.setPrefix("classpath:/test-email-templates/");
+            emailTemplateResolver.setSuffix(".html");
+            emailTemplateResolver.setTemplateMode(TemplateMode.HTML);
+            emailTemplateResolver.setCharacterEncoding(StandardCharsets.UTF_8.name());
+            return emailTemplateResolver;
+        }
+
+        @Bean
+        EmailUtility emailUtility() {
+            return mock(EmailUtility.class);
+        }
+
+        @Bean
         EmailService emailService() {
             return new EmailServiceImpl(
                     this.javaMailSender(),
+                    this.springTemplateEngine(),
+                    this.emailUtility(),
                     this.applicationFrameworkProperties()
             );
         }
@@ -54,6 +90,8 @@ public class EmailServiceImplTest {
 
     // Services
     private final JavaMailSender javaMailSender;
+    // Utilities
+    private final EmailUtility emailUtility;
     // Properties
     private final ApplicationFrameworkProperties applicationFrameworkProperties;
 
@@ -63,6 +101,7 @@ public class EmailServiceImplTest {
     public void beforeEach() {
         reset(
                 this.javaMailSender,
+                this.emailUtility,
                 this.applicationFrameworkProperties
         );
     }
@@ -71,6 +110,7 @@ public class EmailServiceImplTest {
     public void afterEach() {
         verifyNoMoreInteractions(
                 this.javaMailSender,
+                this.emailUtility,
                 this.applicationFrameworkProperties
         );
     }
@@ -81,8 +121,7 @@ public class EmailServiceImplTest {
         var to = randomEmail();
         var subject = randomString();
         var message = randomString();
-        var emailConfigs = new EmailConfigs();
-        emailConfigs.setEnabled(false);
+        var emailConfigs = EmailConfigs.disabled();
         when(this.applicationFrameworkProperties.getEmailConfigs()).thenReturn(emailConfigs);
 
         // Act
@@ -196,5 +235,93 @@ public class EmailServiceImplTest {
         assertThat(simpleMailMessage.getSubject()).isEqualTo(subject);
         assertThat(simpleMailMessage.getText()).isEqualTo(message);
         assertThat(simpleMailMessage.getFrom()).isEqualTo(from);
+    }
+
+    @Test
+    public void sendHTMLDisabledTest() {
+        // Arrange
+        var emailHTML = entity(EmailHTML.class);
+        var emailConfigs = EmailConfigs.disabled();
+        when(this.applicationFrameworkProperties.getEmailConfigs()).thenReturn(emailConfigs);
+
+        // Act
+        this.componentUnderTest.sendHTML(emailHTML);
+
+        // Assert
+        verify(this.applicationFrameworkProperties).getEmailConfigs();
+    }
+
+    @Test
+    public void sendHTMLEnabledExceptionTest() throws MessagingException {
+        // Arrange
+        var from = randomEmail();
+        var emailHTML = entity(EmailHTML.class);
+        var emailConfigs = new EmailConfigs();
+        emailConfigs.setEnabled(true);
+        emailConfigs.setFrom(from);
+        when(this.applicationFrameworkProperties.getEmailConfigs()).thenReturn(emailConfigs);
+        when(this.emailUtility.getMimeMessageTuple2()).thenThrow(new MessagingException());
+
+        // Act
+        this.componentUnderTest.sendHTML(emailHTML);
+
+        // Assert
+        verify(this.applicationFrameworkProperties).getEmailConfigs();
+        verify(this.emailUtility).getMimeMessageTuple2();
+    }
+
+    @Test
+    public void sendHTMLEnabledTest() throws MessagingException {
+        // Arrange
+        var from = randomEmail();
+        Map<String, Object> templateVariables = Map.of(
+                "param1", "key2",
+                "param2", 2L
+        );
+        var emailHTML = EmailHTML.of(
+                Set.of(
+                        "tests@tech1.io"
+                ),
+                "subject1",
+                "template1",
+                templateVariables
+        );
+        var emailConfigs = new EmailConfigs();
+        emailConfigs.setEnabled(true);
+        emailConfigs.setFrom(from);
+        when(this.applicationFrameworkProperties.getEmailConfigs()).thenReturn(emailConfigs);
+        var message = mock(MimeMessage.class);
+        var mimeMessageHelper = mock(MimeMessageHelper.class);
+        when(this.emailUtility.getMimeMessageTuple2()).thenReturn(Tuple2.of(message, mimeMessageHelper));
+
+        // Act
+        this.componentUnderTest.sendHTML(emailHTML);
+
+        // Assert
+        verify(mimeMessageHelper).setFrom(eq(from));
+        verify(mimeMessageHelper).setTo(eq(new String[] { "tests@tech1.io" }));
+        verify(mimeMessageHelper).setSubject(eq("subject1"));
+        var html = "<!DOCTYPE html>\n" +
+                "<html>\n" +
+                "<head>\n" +
+                "  <meta http-equiv=\"Content-Type\" content=\"text/html; charset=UTF-8\"/>\n" +
+                "</head>\n" +
+                "<body>\n" +
+                "<p>\n" +
+                "  Param1: <span >key2</span>\n" +
+                "</p>\n" +
+                "<p>\n" +
+                "  Param2: <span >2</span>\n" +
+                "</p>\n" +
+                "</body>\n" +
+                "</html>\n";
+        verify(mimeMessageHelper).setText(eq(html), eq(true));
+        verify(this.applicationFrameworkProperties).getEmailConfigs();
+        verify(this.emailUtility).getMimeMessageTuple2();
+        verify(this.javaMailSender).send(any(MimeMessage.class));
+        verifyNoMoreInteractions(
+                message,
+                mimeMessageHelper
+        );
     }
 }