+++ RELEASE_NOTES.md
@@ -3,6 +3,7 @@
 ### v1.5
 - Migrate EnumUtility methods, TuplePercentage(long, long)
 - Introduce events/incidents policies
+- Add sendHTML method to EmailService (thymeleaf as template-processor)
 
 ### v1.4
 - Migrate domain-based properties